+++ Net_Core/Web.Angular/projects/gw-lib/core/sys-admin/src/c/db-log-details/db-log-details.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'gw-core-db-log-details',
+  standalone: true,
+  imports: [],
+  templateUrl: './db-log-details.component.html',
+  styleUrl: './db-log-details.component.scss'
+})
+export class DBLogDetailsComponent {
+
+}