+++ Net_Core/Web.Angular/projects/gw-lib/core/sys-admin/src/selected-row.model.ts
@@ -1,0 +1,12 @@
+import { ITotalRecords } from '@growthware/common/interfaces';
+
+export class SelectedRow implements ITotalRecords {
+    Account: string = '';
+    Component: string = '';
+    ClassName: string = '';
+    Level: string = '';
+    LogSeqId: string = '';
+    MethodName: string = '';
+    Msg: string = '';
+    TotalRecords: number = -1;
+}